+++ test/scenario/openstack/run.sh
@@ -32,7 +32,7 @@ function destroy() {
 
   "${__node}" "${__root}/lib/cli.js" destroy \
     --force-yes \
-    --cfg-var="global.terraform.parallelism=1" \
+    --cfg-var="infra.settings.parallelism=1" \
   > /dev/null 2>&1 || true
 }
 
@@ -44,7 +44,7 @@ if true; then trap destroy EXIT; fi
   --no-color \
   --verbose \
   --force-yes \
-  --cfg-var="global.terraform.parallelism=1" \
+  --cfg-var="infra.settings.parallelism=1" \
   --bailAfter install \
 || false
 
@@ -55,7 +55,7 @@ if true; then trap destroy EXIT; fi
 #   --no-color \
 #   --verbose \
 #   --force-yes \
-#   --cfg-var="global.terraform.parallelism=1" \
+#   --cfg-var="infra.settings.parallelism=1" \
 #   --bail \
 # || false